+++ test/fixture/pages/index.vue
@@ -1,5 +1,5 @@
 <template>
-<div>
+  <div>
     /index
-</div>
+  </div>
 </template>